+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Added option to playlist to always enqueue when selected
 - Fixed some edge cases around playlist sync
 - Added catalog slide-out to playlist UI
+- Added move up/down permission and buttons to queue
 - Improved efficiency of playlist UI updates
 
 ### [2.5.0-beta.21]